+++ src/core/ollayerprimitive.js
@@ -28,34 +28,3 @@ olcs.core.OlLayerPrimitive = function(layerProjection, scene) {
   this.add(billboards);
 };
 goog.inherits(olcs.core.OlLayerPrimitive, Cesium.PrimitiveCollection);
-
-
-/**
- * Convert an OpenLayers feature to Cesium primitive collection.
- * @param {!ol.layer.Vector} layer
- * @param {!ol.View} view
- * @param {!ol.Feature} feature
- * @return {Cesium.Primitive}
- * @api
- */
-olcs.core.OlLayerPrimitive.prototype.convert = function(layer, view, feature) {
-  var proj = view.getProjection();
-  var resolution = view.getResolution();
-
-  if (!goog.isDef(resolution) || !proj) {
-    return null;
-  }
-
-  var layerStyle = layer.getStyleFunction();
-  var style = olcs.core.computePlainStyle(feature, layerStyle, resolution);
-
-  if (!style) {
-    // only 'render' features with a style
-    return null;
-  }
-
-  this.context.projection = proj;
-  return olcs.core.olFeatureToCesium(feature, style, this.context);
-};
-
-
